Libs/requirements.txt
pyserial
pymongo
python-dotenv
requests